+++ src/datatypes/vec3_2.c
@@ -6,7 +6,7 @@
 /*   By: jcat <joaoteix@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/02 01:15:03 by jcat              #+#    #+#             */
-/*   Updated: 2024/04/07 02:25:28 by jcat             ###   ########.fr       */
+/*   Updated: 2024/05/05 16:42:24 by joaoteix         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ t_vec3	v3sub(const t_vec3 a, const t_vec3 b)
 	return ((t_vec3){a.x - b.x, a.y - b.y, a.z - b.z});
 }
 
-t_vec3	vec3(const float x, const float y, const float z)
+t_vec3	vec3(const double x, const double y, const double z)
 {
 	return ((t_vec3){x, y, z});
 }
